requirements.txt
Adafruit-Blinka==2.3.2
adafruit-circuitpython-neopixel==3.3.7
Adafruit-PlatformDetect==1.3.2
Adafruit-PureIO==0.2.3
pkg-resources==0.0.0
rpi-ws281x==4.2.2
RPi.GPIO==0.7.0
spidev==3.4
sysv-ipc==1.0.0
